+++ betfairlightweight/endpoints/logout.py
@@ -6,7 +6,7 @@ from .baseendpoint import BaseEndpoint
 from ..resources import LogoutResource
 from ..exceptions import LogoutError, APIError, InvalidResponse
 from ..utils import check_status_code
-from ..compat import json_loads
+from ..compat import json
 
 
 class Logout(BaseEndpoint):
@@ -48,7 +48,7 @@ class Logout(BaseEndpoint):
 
         check_status_code(response)
         try:
-            response_json = json_loads(response.content.decode("utf-8"))
+            response_json = json.loads(response.content.decode("utf-8"))
         except ValueError:
             raise InvalidResponse(response.text)