+++ betfairlightweight/filters.py
@@ -1,3 +1,5 @@
+import datetime
+from typing import Union
 from .utils import to_camel_case
 
 from .resources import bettingresources
@@ -65,13 +67,29 @@ def streaming_order_filter(
     return {to_camel_case(k): v for k, v in args.items() if v is not None}
 
 
-def time_range(from_: str = None, to: str = None) -> dict:  # todo datetime conversion
+def time_range(
+    from_: Union[str, datetime.datetime] = None,
+    to: Union[str, datetime.datetime] = None,
+) -> dict:
     """
-    :param str from_:
-    :param str to:
+    :param Union[str, datetime.datetime] from_:
+    :param Union[str, datetime.datetime] to:
 
     :return: dict
     """
+
+    if from_ != None:
+        if isinstance(from_, datetime.datetime):
+            from_ = from_.isoformat()
+        elif not isinstance(from_, str):
+            raise TypeError("The 'from_' value must be string or datetime (not date)")
+
+    if to != None:
+        if isinstance(to, datetime.datetime):
+            to = to.isoformat()
+        elif not isinstance(to, str):
+            raise TypeError("The 'to' value must be string or datetime (not date)")
+
     args = locals().copy()
     return {k.replace("_", ""): v for k, v in args.items()}