+++ HISTORY.rst
@@ -3,6 +3,31 @@
 Release History
 ---------------
 
+2.2.0 (2020-03-09)
++++++++++++++++++++
+
+**Improvements**
+
+- #283 max_latency can now be set to None
+
+**Dependencies**
+
+- requests upgraded from 2.22.0 to 2.23.0
+- ujson upgraded from 1.35 to 2.0.1 (updates to compat.py)
+
+2.1.0 (2020-03-02)
++++++++++++++++++++
+
+**Improvements**
+
+- datetime handling added to time_range filter (@trigvi)
+- connectionsAvailable handling added
+
+**Bug Fixes**
+
+- #273 error handling added for markets without marketDefinition
+- #233 sendall used instead of send so that all data is sent (bug present since 2016!)
+
 2.0.1 (2020-02-17)
 +++++++++++++++++++